# Performance optimizations preprocessing callgrind data